# perf: optimize FunctionCallbackInfo and PropertyCallbackInfo methods