### Before:

```
Running function_overhead ...
  15.1 ns per run 66.2 million ops/sec → new_
  15.2 ns per run 65.7 million ops/sec → new_raw
  11.2 ns per run 89.7 million ops/sec → new_set_uint32
  6.8 ns per run 147.1 million ops/sec → new_raw_set_uint32
  2.7 ns per run 367.6 million ops/sec → new_fast
Running primitives ...
  11.7 ns per run 85.3 million ops/sec → undefined_from_scope
  7.5 ns per run 133.2 million ops/sec → undefined_from_isolate
```

### After:

```
Running function_overhead ...
  14.6 ns per run 68.7 million ops/sec → new_
  14.7 ns per run 67.9 million ops/sec → new_raw
  10.9 ns per run 91.7 million ops/sec → new_set_uint32
  6.8 ns per run 147.9 million ops/sec → new_raw_set_uint32
  2.7 ns per run 367.6 million ops/sec → new_fast
Running primitives ...
  11.2 ns per run 89.5 million ops/sec → undefined_from_scope
  6.3 ns per run 158.7 million ops/sec → undefined_from_isolate
```